% Generated by roxygen2 (4.0.2): do not edit by hand
\name{bdt}
\alias{bdt}
\title{Black-Derman-Toy Interest Rate Tree}
\usage{
bdt(yields, volatilities)
}
\arguments{
\item{yields}{The historical zero-coupon bond yields}

\item{volatilities}{The volatilities in force for each of the zero-coupon bonds at time 1. Note that volatilites[1] is undefined}
}
\description{
Constructs the short rate tree using the data calibration model proposed by Black and Derman
}
\details{
uses a version of newtons method that generalizes to n dimensions. May not converge for particular historical data
}
\examples{
bdt(yields=c(0.10, 0.11, 0.12, 0.125), volatilities=c(NA, 0.10, 0.15, 0.14))
}

