% Generated by roxygen2 (4.0.2): do not edit by hand
\name{currentStocks}
\alias{currentStocks}
\title{Current Stock Info (Yahoo Finance API)}
\usage{
currentStocks(stockIndex = "GOOGL")
}
\arguments{
\item{stockIndex}{The name of the stock index to retrieve information for (default GOOGL)}
}
\description{
Parses the Yahoo Finance API to retrieve current stock information
}
\details{
Uses R's read.csv method and generates a data table containing stock information such as Bid-Ask, EPS.
}
\examples{
currentStocks()
}

