% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Gamma_c}
\alias{Gamma_c}
\title{Gamma (Call Greek)}
\usage{
Gamma_c(S, K, sigma, r, delta, t)
}
\arguments{
\item{S}{The Stock Price}

\item{K}{The Strike Price}

\item{sigma}{The volatility}

\item{r}{The continuously compounded risk-tree interest rate}

\item{delta}{The annualized dividend rate}

\item{t}{The expiration date}
}
\description{
A partial derivative of the Black-Scholes Equation: dc^2/dS^2 (with respect to the stock price)
}
\examples{
Gamma_c(S=40,K=40,sigma=0.3,r=0.08,delta=0,t=91/365)
}

