% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Rho_c}
\alias{Rho_c}
\title{Rho (Call Greek)}
\usage{
Rho_c(S, K, sigma, r, delta, t)
}
\arguments{
\item{S}{The Stock Price}

\item{K}{The Strike Price}

\item{sigma}{The volatility}

\item{r}{The continuously compounded risk-tree interest rate}

\item{delta}{The annualized dividend rate}

\item{t}{The expiration date}
}
\description{
A partial derivative of the Black-Scholes Equation: dc/dr (with respect to the interest rate)
}
\examples{
Rho_c(S=40,K=40,sigma=0.3,r=0.08,delta=0,t=91/365)
}

