% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treeBasic}
\alias{treeBasic}
\title{European Option Price (Basic Binomial Tree Formula)}
\usage{
treeBasic(n = 3, call = TRUE)
}
\arguments{
\item{n}{the number of periods}

\item{call}{TRUE if call option, FALSE if put option}
}
\description{
Uses Cox's Formula for finding the price of a European call option.
}
\details{
This formula assumes that nodes do overlap and path dependency is not needed as in an Asian option.
}
\examples{
treeBasic(n=10, call=TRUE)
}

