% Generated by roxygen2 (4.0.2): do not edit by hand
\name{treeDetails}
\alias{treeDetails}
\title{European Option Pricing Details}
\usage{
treeDetails(n = 2, call = TRUE)
}
\arguments{
\item{n}{the number of periods}

\item{call}{TRUE if call option, FALSE if put option}
}
\description{
Returns a tree structure with the D (Delta: amount of shares to buy), B (lend: borrowing amount), S (stock price), C (option value) at each node
}
\details{
Uses a list structure to store the data. If n is large, the object is very large and takes up a lot of memory
}
\examples{
treeDetails(n=10, call=TRUE)
}

